pyproject.toml
[tool.poetry]
name = "themis-llm"
version = "0.1.0"
description = "Themis"
authors = ["Panagiotis Papadakos <papadako@gmail.com>, Christos Karanikolopoulos <chkaranikolopoulos@cs.uoi.gr>"]
readme = "README.md"
packages = [{include = "themis"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.3"
pandas = "^2.2.2"
matplotlib = "^3.9.1"
vllm = "^0.5.1"
huggingface-hub = {extras = ["cli"], version = "^0.23.4"}
wandb = "^0.17.4"
langchain-community = "^0.2.7"
lm-eval = "^0.4.4"
python-json-logger = "^2.0.7"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
chardet = "^5.2.0"
black = "^24.4.2"
pylint = "^3.2.5"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
